Move einsum testing utilities to funsor.testing
